refactor(application): rename class
