+++ app/src/test/java/fr/carbonit/th/simulation/TreasureMapTest.java
@@ -1,54 +1,54 @@
 package fr.carbonit.th.simulation;
 
-import fr.carbonit.th.command.Coordinates;
 import fr.carbonit.th.configuration.TreasureHuntConfiguration;
+import fr.carbonit.th.provider.ConfigurationProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class SimulatorTest {
+public class TreasureMapTest {
 
     private final ConfigurationProvider provider;
-    private Simulator simulator;
-
-    public SimulatorTest() {
+    private TreasureMap map;
+    
+    public TreasureMapTest() {
         provider = new ConfigurationProvider();
     }
 
     @BeforeEach
     public void init() {
         TreasureHuntConfiguration configuration = provider.provide();
-        simulator = new Simulator(configuration);
+        map = new TreasureMap(configuration);
     }
 
     @Test
     public void shouldHaveAMapOfCoordinatesAsProperty() {
-        assertNotEquals(null, simulator.getMap());
+        assertNotEquals(null, map.getCells());
     }
 
     @Test
-    public void shouldHaveASizeMapEqualsTo12() {
-        assertEquals(12, simulator.getMap().size());
+    public void shouldHaveAnAreaEqualsTo12() {
+        assertEquals(12, map.getCells().size());
     }
 
     @Test
     public void shouldFirstMapKeyEqualsTo0_0() {
-        assertTrue(simulator.getMap().containsKey(new Coordinates(0, 0)));
+        assertTrue(map.getCells().containsKey(new Coordinates(0, 0)));
     }
 
     @Test
     public void shouldLastKeyMapKeyEqualsTo2_3() {
-        assertTrue(simulator.getMap().containsKey(new Coordinates(2, 3)));
+        assertTrue(map.getCells().containsKey(new Coordinates(2, 3)));
     }
 
     @Test
     public void shouldNotContainCoordinates3_3() {
-        assertFalse(simulator.getMap().containsKey(new Coordinates(3, 3)));
+        assertFalse(map.getCells().containsKey(new Coordinates(3, 3)));
     }
 
     @Test
     public void shouldNotContainCoordinates2_4() {
-        assertFalse(simulator.getMap().containsKey(new Coordinates(2, 4)));
+        assertFalse(map.getCells().containsKey(new Coordinates(2, 4)));
     }
 }
